+++ resources/js/index.js
@@ -1,0 +1,5 @@
+import { defineElement } from './utils.js'
+import { Sidebar, SidebarToggle } from './sidebar.js'
+
+defineElement("sidebar", Sidebar);
+defineElement("sidebar-toggle", SidebarToggle);